#230 phase-management change endPoint get phases/1/iterations for get all phase iterations

+++ rup-projects-core/src/test/java/com/core/rupprojectscore/service/ProjectServiceTest.java
@@ -131,7 +131,7 @@ public class ProjectServiceTest {
 
     private void assertIteration(Long order, IterationDto iteration, List<? extends Serializable>[] expectedIterationInfo) {
         assertThat(Arrays.asList(iteration.getStartDate(), iteration.getEndDate())).isIn(expectedIterationInfo);
-        assertThat(iteration.getOrder()).isEqualTo(order);
+        assertThat(iteration.getOrder()).isEqualTo((long) order);
     }
 
     private long cost(long cost) {
